==> educacao_prefeituras_idados/03 - escolas_nordeste.R <==
# Dados de educacao 2020 - Base ESCOLAS
# modelo da consultoria iDados
# Renato Gomes Chaves
# Criacao 08/10/2021
# Ultima modificacao: 22/11/2021

# Carregar Pacotes #############################################################
{
  require(pacman)
  p_load(tidyverse,
         here
  )
  # no_cores <- detectCores()
  # registerDoParallel(no_cores)
  # aumentar max mem ram com pagefile
  # utils::memory.limit(32000)
  # cat('Memoria RAM maxima:', memory.limit())
  # options(future.globals.maxSize = 2500 * 1024^2)
}

# Carregar dados ###############################################################
# Carregar dados das regioes brasileiras
regioes_br <- readxl::read_xlsx(here('data/extra/regioes_geograficas_composicao_por_municipios_2017.xlsx'),
                                col_types = 'text')

# Subdividir regioes
# por Estados
estados_br <- regioes_br %>%
  group_by(nome_uf,
           co_uf = cod_uf) %>%
  summarise() %>% 
  mutate(nome_uf = as_factor(nome_uf))

# por Municipios
municipios_br <- regioes_br %>%
  group_by(nome_mun,
           co_municipio_dv = cod_mun_dv) %>%
  summarise() %>% 
  mutate(nome_mun = as_factor(nome_mun))

# Carregar apenas as colunas selecionadas
escolas_2020_nordeste <- read_delim(here('data/raw/microdados_educacao_basica_2020/DADOS/escolas.CSV'),
                                   '|',
                                   # n_max = 10000,
                                   # escape_double = FALSE,
                                   trim_ws = TRUE,
                                   locale = locale(encoding = "ISO-8859-1",
                                                   decimal_mark = ','),
                                   col_types = cols_only(# DADOS DAS ESCOLAS
                                     # NU_ANO_CENSO = col_double(),
                                     CO_ENTIDADE = 'c',
                                     NO_ENTIDADE = 'c',
                                     # CO_ORGAO_REGIONAL = col_character(),
                                     # TP_SITUACAO_FUNCIONAMENTO = col_double(),
                                     # DT_ANO_LETIVO_INICIO = col_character(),
                                     # DT_ANO_LETIVO_TERMINO = col_character(),
                                     # CO_REGIAO = col_double(),
                                     # CO_MESORREGIAO = col_double(),
                                     # CO_MICRORREGIAO = col_double(),
                                     CO_UF = 'c',
                                     CO_MUNICIPIO = 'c',
                                     # CO_DISTRITO = col_double(),
                                     # TP_DEPENDENCIA = col_double(),
                                     # TP_LOCALIZACAO = col_double(),
                                     # TP_LOCALIZACAO_DIFERENCIADA = col_double(),
                                     # IN_VINCULO_SECRETARIA_EDUCACAO = col_double(),
                                     # IN_VINCULO_SEGURANCA_PUBLICA = col_double(),
                                     # IN_VINCULO_SECRETARIA_SAUDE = col_double(),
                                     # IN_VINCULO_OUTRO_ORGAO = col_double(),
                                     # TP_CATEGORIA_ESCOLA_PRIVADA = col_double(),
                                     # IN_CONVENIADA_PP = col_double(),
                                     # TP_CONVENIO_PODER_PUBLICO = col_double(),
                                     # IN_MANT_ESCOLA_PRIVADA_EMP = col_double(),
                                     # IN_MANT_ESCOLA_PRIVADA_ONG = col_double(),
                                     # IN_MANT_ESCOLA_PRIVADA_OSCIP = col_double(),
                                     # IN_MANT_ESCOLA_PRIV_ONG_OSCIP = col_double(),
                                     # IN_MANT_ESCOLA_PRIVADA_SIND = col_double(),
                                     # IN_MANT_ESCOLA_PRIVADA_SIST_S = col_double(),
                                     # IN_MANT_ESCOLA_PRIVADA_S_FINS = col_double(),
                                     # TP_REGULAMENTACAO = col_double(),
                                     # TP_RESPONSAVEL_REGULAMENTACAO = col_double(),
                                     # CO_ESCOLA_SEDE_VINCULADA = col_double(),
                                     # CO_IES_OFERTANTE = col_double(),
                                     # IN_LOCAL_FUNC_PREDIO_ESCOLAR = col_double(),
                                     # TP_OCUPACAO_PREDIO_ESCOLAR = col_double(),
                                     # IN_LOCAL_FUNC_SOCIOEDUCATIVO = col_double(),
                                     # IN_LOCAL_FUNC_UNID_PRISIONAL = col_double(),
                                     # IN_LOCAL_FUNC_PRISIONAL_SOCIO = col_double(),
                                     # IN_LOCAL_FUNC_GALPAO = col_double(),
                                     # TP_OCUPACAO_GALPAO = col_double(),
                                     # IN_LOCAL_FUNC_SALAS_OUTRA_ESC = col_double(),
                                     # IN_LOCAL_FUNC_OUTROS = col_double(),
                                     # IN_PREDIO_COMPARTILHADO = col_double(),
                                     # IN_AGUA_POTAVEL = col_double(),
                                     # IN_AGUA_REDE_PUBLICA = col_double(),
                                     # IN_AGUA_POCO_ARTESIANO = col_double(),
                                     # IN_AGUA_CACIMBA = col_double(),
                                     # IN_AGUA_FONTE_RIO = col_double(),
                                     # IN_AGUA_INEXISTENTE = col_double(),
                                     # IN_ENERGIA_REDE_PUBLICA = col_double(),
                                     # IN_ENERGIA_GERADOR_FOSSIL = col_double(),
                                     # IN_ENERGIA_RENOVAVEL = col_double(),
                                     # IN_ENERGIA_INEXISTENTE = col_double(),
                                     # IN_ESGOTO_REDE_PUBLICA = col_double(),
                                     # IN_ESGOTO_FOSSA_SEPTICA = col_double(),
                                     # IN_ESGOTO_FOSSA_COMUM = col_double(),
                                     # IN_ESGOTO_FOSSA = col_double(),
                                     # IN_ESGOTO_INEXISTENTE = col_double(),
                                     # IN_LIXO_SERVICO_COLETA = col_double(),
                                     # IN_LIXO_QUEIMA = col_double(),
                                     # IN_LIXO_ENTERRA = col_double(),
                                     # IN_LIXO_DESTINO_FINAL_PUBLICO = col_double(),
                                     # IN_LIXO_DESCARTA_OUTRA_AREA = col_double(),
                                     # IN_TRATAMENTO_LIXO_SEPARACAO = col_double(),
                                     # IN_TRATAMENTO_LIXO_REUTILIZA = col_double(),
                                     # IN_TRATAMENTO_LIXO_RECICLAGEM = col_double(),
                                     # IN_TRATAMENTO_LIXO_INEXISTENTE = col_double(),
                                     # IN_ALMOXARIFADO = col_double(),
                                     # IN_AREA_VERDE = col_double(),
                                     # IN_AUDITORIO = col_double(),
                                     # IN_BANHEIRO = col_double(),
                                     # IN_BANHEIRO_EI = col_double(),
                                     # IN_BANHEIRO_PNE = col_double(),
                                     # IN_BANHEIRO_FUNCIONARIOS = col_double(),
                                     # IN_BANHEIRO_CHUVEIRO = col_double(),
                                     # IN_BIBLIOTECA = col_double(),
                                     # IN_BIBLIOTECA_SALA_LEITURA = col_double(),
                                     # IN_COZINHA = col_double(),
                                     # IN_DESPENSA = col_double(),
                                     # IN_DORMITORIO_ALUNO = col_double(),
                                     # IN_DORMITORIO_PROFESSOR = col_double(),
                                     # IN_LABORATORIO_CIENCIAS = col_double(),
                                     # IN_LABORATORIO_INFORMATICA = col_double(),
                                     # IN_PATIO_COBERTO = col_double(),
                                     # IN_PATIO_DESCOBERTO = col_double(),
                                     # IN_PARQUE_INFANTIL = col_double(),
                                     # IN_PISCINA = col_double(),
                                     # IN_QUADRA_ESPORTES = col_double(),
                                     # IN_QUADRA_ESPORTES_COBERTA = col_double(),
                                     # IN_QUADRA_ESPORTES_DESCOBERTA = col_double(),
                                     # IN_REFEITORIO = col_double(),
                                     # IN_SALA_ATELIE_ARTES = col_double(),
                                     # IN_SALA_MUSICA_CORAL = col_double(),
                                     # IN_SALA_ESTUDIO_DANCA = col_double(),
                                     # IN_SALA_MULTIUSO = col_double(),
                                     # IN_SALA_DIRETORIA = col_double(),
                                     # IN_SALA_LEITURA = col_double(),
                                     # IN_SALA_PROFESSOR = col_double(),
                                     # IN_SALA_REPOUSO_ALUNO = col_double(),
                                     # IN_SECRETARIA = col_double(),
                                     # IN_SALA_ATENDIMENTO_ESPECIAL = col_double(),
                                     # IN_TERREIRAO = col_double(),
                                     # IN_VIVEIRO = col_double(),
                                     # IN_DEPENDENCIAS_OUTRAS = col_double(),
                                     # IN_ACESSIBILIDADE_CORRIMAO = col_double(),
                                     # IN_ACESSIBILIDADE_ELEVADOR = col_double(),
                                     # IN_ACESSIBILIDADE_PISOS_TATEIS = col_double(),
                                     # IN_ACESSIBILIDADE_VAO_LIVRE = col_double(),
                                     # IN_ACESSIBILIDADE_RAMPAS = col_double(),
                                     # IN_ACESSIBILIDADE_SINAL_SONORO = col_double(),
                                     # IN_ACESSIBILIDADE_SINAL_TATIL = col_double(),
                                     # IN_ACESSIBILIDADE_SINAL_VISUAL = col_double(),
                                     # IN_ACESSIBILIDADE_INEXISTENTE = col_double(),
                                     # QT_SALAS_UTILIZADAS_DENTRO = col_double(),
                                     # QT_SALAS_UTILIZADAS_FORA = col_double(),
                                     QT_SALAS_UTILIZADAS = 'd',
                                     # QT_SALAS_UTILIZA_CLIMATIZADAS = col_double(),
                                     # QT_SALAS_UTILIZADAS_ACESSIVEIS = col_double(),
                                     # IN_EQUIP_PARABOLICA = col_double(),
                                     # IN_COMPUTADOR = col_double(),
                                     # IN_EQUIP_COPIADORA = col_double(),
                                     # IN_EQUIP_IMPRESSORA = col_double(),
                                     # IN_EQUIP_IMPRESSORA_MULT = col_double(),
                                     # IN_EQUIP_SCANNER = col_double(),
                                     # IN_EQUIP_NENHUM = col_double(),
                                     # IN_EQUIP_DVD = col_double(),
                                     # QT_EQUIP_DVD = col_double(),
                                     # IN_EQUIP_SOM = col_double(),
                                     # QT_EQUIP_SOM = col_double(),
                                     # IN_EQUIP_TV = col_double(),
                                     # QT_EQUIP_TV = col_double(),
                                     # IN_EQUIP_LOUSA_DIGITAL = col_double(),
                                     # QT_EQUIP_LOUSA_DIGITAL = col_double(),
                                     # IN_EQUIP_MULTIMIDIA = col_double(),
                                     # QT_EQUIP_MULTIMIDIA = col_double(),
                                     # IN_DESKTOP_ALUNO = col_double(),
                                     # QT_DESKTOP_ALUNO = col_double(),
                                     # IN_COMP_PORTATIL_ALUNO = col_double(),
                                     # QT_COMP_PORTATIL_ALUNO = col_double(),
                                     # IN_TABLET_ALUNO = col_double(),
                                     # QT_TABLET_ALUNO = col_double(),
                                     # IN_INTERNET = col_double(),
                                     # IN_INTERNET_ALUNOS = col_double(),
                                     # IN_INTERNET_ADMINISTRATIVO = col_double(),
                                     # IN_INTERNET_APRENDIZAGEM = col_double(),
                                     # IN_INTERNET_COMUNIDADE = col_double(),
                                     # IN_ACESSO_INTERNET_COMPUTADOR = col_double(),
                                     # IN_ACES_INTERNET_DISP_PESSOAIS = col_double(),
                                     # TP_REDE_LOCAL = col_double(),
                                     # IN_BANDA_LARGA = col_double(),
                                     # QT_PROF_ADMINISTRATIVOS = col_double(),
                                     # QT_PROF_SERVICOS_GERAIS = col_double(),
                                     # QT_PROF_BIBLIOTECARIO = col_double(),
                                     # QT_PROF_SAUDE = col_double(),
                                     # QT_PROF_COORDENADOR = col_double(),
                                     # QT_PROF_FONAUDIOLOGO = col_double(),
                                     # QT_PROF_NUTRICIONISTA = col_double(),
                                     # QT_PROF_PSICOLOGO = col_double(),
                                     # QT_PROF_ALIMENTACAO = col_double(),
                                     # QT_PROF_PEDAGOGIA = col_double(),
                                     # QT_PROF_SECRETARIO = col_double(),
                                     # QT_PROF_SEGURANCA = col_double(),
                                     # QT_PROF_MONITORES = col_double(),
                                     # QT_PROF_GESTAO = col_double(),
                                     # QT_PROF_ASSIST_SOCIAL = col_double(),
                                     # IN_ALIMENTACAO = col_double(),
                                     # IN_SERIE_ANO = col_double(),
                                     # IN_PERIODOS_SEMESTRAIS = col_double(),
                                     # IN_FUNDAMENTAL_CICLOS = col_double(),
                                     # IN_GRUPOS_NAO_SERIADOS = col_double(),
                                     # IN_MODULOS = col_double(),
                                     # IN_FORMACAO_ALTERNANCIA = col_double(),
                                     # IN_MATERIAL_PED_MULTIMIDIA = col_double(),
                                     # IN_MATERIAL_PED_INFANTIL = col_double(),
                                     # IN_MATERIAL_PED_CIENTIFICO = col_double(),
                                     # IN_MATERIAL_PED_DIFUSAO = col_double(),
                                     # IN_MATERIAL_PED_MUSICAL = col_double(),
                                     # IN_MATERIAL_PED_JOGOS = col_double(),
                                     # IN_MATERIAL_PED_ARTISTICAS = col_double(),
                                     # IN_MATERIAL_PED_DESPORTIVA = col_double(),
                                     # IN_MATERIAL_PED_INDIGENA = col_double(),
                                     # IN_MATERIAL_PED_ETNICO = col_double(),
                                     # IN_MATERIAL_PED_CAMPO = col_double(),
                                     # IN_MATERIAL_PED_NENHUM = col_double(),
                                     # IN_EDUCACAO_INDIGENA = col_double(),
                                     # TP_INDIGENA_LINGUA = col_double(),
                                     # CO_LINGUA_INDIGENA_1 = col_double(),
                                     # CO_LINGUA_INDIGENA_2 = col_double(),
                                     # CO_LINGUA_INDIGENA_3 = col_double(),
                                     # IN_EXAME_SELECAO = col_double(),
                                     # IN_RESERVA_PPI = col_double(),
                                     # IN_RESERVA_RENDA = col_double(),
                                     # IN_RESERVA_PUBLICA = col_double(),
                                     # IN_RESERVA_PCD = col_double(),
                                     # IN_RESERVA_OUTROS = col_double(),
                                     # IN_RESERVA_NENHUMA = col_double(),
                                     # IN_REDES_SOCIAIS = col_double(),
                                     # IN_ESPACO_ATIVIDADE = col_double(),
                                     # IN_ESPACO_EQUIPAMENTO = col_double(),
                                     # IN_ORGAO_ASS_PAIS = col_double(),
                                     # IN_ORGAO_ASS_PAIS_MESTRES = col_double(),
                                     # IN_ORGAO_CONSELHO_ESCOLAR = col_double(),
                                     # IN_ORGAO_GREMIO_ESTUDANTIL = col_double(),
                                     # IN_ORGAO_OUTROS = col_double(),
                                     # IN_ORGAO_NENHUM = col_double(),
                                     # TP_PROPOSTA_PEDAGOGICA = col_double(),
#################################### DADOS DA OFERTA DE MATRICULA
                                     # TP_AEE = col_double(),
                                     # TP_ATIVIDADE_COMPLEMENTAR = col_double(),
                                     # IN_MEDIACAO_PRESENCIAL = col_double(),
                                     # IN_MEDIACAO_SEMIPRESENCIAL = col_double(),
                                     # IN_MEDIACAO_EAD = col_double(),
                                     # IN_ESPECIAL_EXCLUSIVA = col_double(),
                                     # IN_REGULAR = col_double(),
                                     # IN_EJA = col_double(),
                                     # IN_PROFISSIONALIZANTE = col_double(),
                                     # IN_COMUM_CRECHE = col_double(),
                                     # IN_COMUM_PRE = col_double(),
                                     # IN_COMUM_FUND_AI = col_double(),
                                     # IN_COMUM_FUND_AF = col_double(),
                                     # IN_COMUM_MEDIO_MEDIO = col_double(),
                                     # IN_COMUM_MEDIO_INTEGRADO = col_double(),
                                     # IN_COMUM_MEDIO_NORMAL = col_double(),
                                     # IN_ESP_EXCLUSIVA_CRECHE = col_double(),
                                     # IN_ESP_EXCLUSIVA_PRE = col_double(),
                                     # IN_ESP_EXCLUSIVA_FUND_AI = col_double(),
                                     # IN_ESP_EXCLUSIVA_FUND_AF = col_double(),
                                     # IN_ESP_EXCLUSIVA_MEDIO_MEDIO = col_double(),
                                     # IN_ESP_EXCLUSIVA_MEDIO_INTEGR = col_double(),
                                     # IN_ESP_EXCLUSIVA_MEDIO_NORMAL = col_double(),
                                     # IN_COMUM_EJA_FUND = col_double(),
                                     # IN_COMUM_EJA_MEDIO = col_double(),
                                     # IN_COMUM_EJA_PROF = col_double(),
                                     # IN_ESP_EXCLUSIVA_EJA_FUND = col_double(),
                                     # IN_ESP_EXCLUSIVA_EJA_MEDIO = col_double(),
                                     # IN_ESP_EXCLUSIVA_EJA_PROF = col_double(),
                                     # IN_COMUM_PROF = col_double(),
                                     # IN_ESP_EXCLUSIVA_PROF = col_double()
                                   )
) %>% 
  left_join(estados_br, by = c('CO_UF' = 'co_uf')) %>% 
  left_join(municipios_br, by = c('CO_MUNICIPIO' = 'co_municipio_dv'))

# Salvar DB resultante #########################################################
write_rds(escolas_2020_nordeste, here('data/processed/escolas_2020_nordeste.rds'))
